Fix Dashboard message inconsistency
The sync() method in HdmiDisplayListener missed resetting it's state,
resulting in falsely reporting that a display was connected.

This showed up if the phone display was powered off while disconnected
from an external display. When the Dashboard was opened again by
unlocking the screen, sync() would trigger, but would still report a
connected display due to the stale display id.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>