maru: Update Desktop Settings for N
There were some decent changes in N with the way Settings works.
Settings dashboard categories need to be declared in
AndroidManifest.xml, fragments need to have the right category tags in
AndroidManifest.xml, and tiles not included in the restricted users
array must be explicitly added based on whether the current user is an
admin or not.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>