Add desktop settings
These patches are heavily based on the old MDesktopSettings used in Maru
on Android L, M, and N.

The main difference is that we switch to extending DashboardFragment and use
a Preference item instead of a custom layout.

This also sets the stage for introducing additional preference items to
manage Maru Desktop in the future.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>